Refactor; extract make areas step. Split up steps to try to understand what the actually do.
